% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCriteria.R
\name{getCriteria}
\alias{getCriteria}
\title{get criteria used in tasting}
\usage{
getCriteria(
  host = "https://topdegust.trydea.fr/api/topdegust",
  token,
  id_tasting = NULL
)
}
\arguments{
\item{host}{url of topdegust API, by default "https://topdegust.trydea.fr/api/topdegust"}

\item{token}{a token from topdegust API}

\item{id_tasting}{tasting id}
}
\value{
a data.frame with a row for each criteria and id, name, lower_bound,upper_bound, description,deleted,categorie,type...as cols
}
\description{
this function provides all criteria used in topdegust, or only criteria used in a given tasting.
If param id_degust is left empty, the function returns all criteria used in topdegust.
}
