% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNotes.R
\name{getNotes}
\alias{getNotes}
\title{get notes from tasting}
\usage{
getNotes(
  host = "https://topdegust.trydea.fr/api/topdegust",
  token,
  id_tasting = NULL
)
}
\arguments{
\item{host}{url of topdegust API, by default "https://topdegust.trydea.fr/api/topdegust"}

\item{token}{a token from topdegust API}

\item{id_tasting}{tasting id}
}
\value{
a data.frama with a row for each combination jugde*product and tasting criteria as cols
}
\description{
get notes from tasting
}
