% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTasting.R
\name{getTasting}
\alias{getTasting}
\title{get metadata from tasting}
\usage{
getTasting(
  host = "https://topdegust.trydea.fr/api/topdegust",
  token,
  year = NULL,
  type = NULL,
  id_tasting = NULL,
  user = NULL
)
}
\arguments{
\item{host}{url of topdegust API, by default "https://topdegust.trydea.fr/api/topdegust"}

\item{token}{a token from topdegust API}

\item{year}{year of the tasting, numeric}

\item{type}{type of the tasting, character}

\item{id_tasting}{tasting id, numeric}

\item{user}{user identifier, character}
}
\value{
a data.frama with a row for each tasting and metadata as cols
}
\description{
get metadata from tasting
}
